+++ CHANGELOG.md
@@ -1,6 +1,10 @@
+# 1.0.2 (unreleased)
+
+* Add `FRACTAL_RUNNER_MAX_TASKS_PER_WORKFLOW` configuration variable (\#469).
+
 # 1.0.1
 
-* Fix bug with environment variable names (468).
+* Fix bug with environment variable names (\#468).
 
 # 1.0.0